+++ src/jumplist.c
@@ -6,8 +6,6 @@ extern "C" {
 
 #if CYGWIN_VERSION_API_MINOR >= 74
 
-//#undef WINVER
-//#define WINVER 0x0601
 //#undef NTDDI_VERSION
 //#define NTDDI_VERSION NTDDI_WIN7
 #ifndef ___PROCESSOR_NUMBER_DEFINED
@@ -19,6 +17,8 @@ typedef struct _PROCESSOR_NUMBER {
 } PROCESSOR_NUMBER, *PPROCESSOR_NUMBER;
 #endif
 
+//#undef WINVER
+//#define WINVER 0x0601
 #include <shlobj.h>
 #include <propkey.h>  // PKEY_Title
 
@@ -245,7 +245,7 @@ HRESULT
 setup_jumplist(wstring appid, int n, wstring titles[], wstring cmds[], wstring icons[], int ii[])
 {
   (void)appid; (void)n; (void)titles; (void)cmds; (void)icons; (void)ii;
-  return S_OK;
+  return 0;
 }
 
 #endif