+++ src/term.h
@@ -657,6 +657,7 @@ struct term {
   uchar vt52_mode;
   bool save_autowrap;
   bool save_rev_wrap;
+  bool emoji_width;
 
   uint csi_argc;
   uint csi_argv[32];