+++ wiki/Changelog.md
@@ -1,11 +1,19 @@
+Terminal features
+  * Emoji width mode (DECSET 2027) to always render emojis in 2-cell width.
+  * Dropped deprecated DECSET mode 2027 for terminal reflow.
+
 Character encoding
   * Disable GB18030 workaround for cygwin since 3.5 which supports it natively.
+
+Font rendering
+  * Fix and enable rendering for geometric Powerline symbols (#979, #943).
 
 Mouse and menu handling
   * Allow disabling of mouse menus by clearing Menu* (#1279).
 
 Configuration
   * Setting options Menu* to empty value disables the respective menu (#1279).
+  * Skip config files in $HOME if $HOME does not exist (mintty/wsltty#351).
 
 ### 3.7.4 (03 July 2024) ###